complete 'destroy' and 'update' in pssword checking

+++ app/controllers/api/v1/users_controller.rb
@@ -29,33 +29,38 @@ module Api
 	 	  	#   "todos": "[{'is_done':'false',title:'Chicken wings3',expire:null},{is_done:false,title:'Scotch eggs2',expire:null},{is_done:'false','title':'Tiramisù2','expire':'null'}]",
    			#   "is_all_done": "false",
    			#   "user_name": "rachel_green2",
-			# 	"password": "12345"
-			#   "password_confirmation": "12345"
+			  # 	"password": "12345"
+			  #   "password_confirmation": "12345"
 			# }
 
 			def create
 				user = User.new(user_params)
-			
-				if user.save
-					render json: {status: 'SUCCESS', message:'Saved shoppinglist', data:user},status: :ok
+				if !(User.find_by_user_name(user.user_name))
+					if user.save
+						render json: {status: 'SUCCESS', message:'Saved shoppinglist', data:user},status: :ok
+	 				else
+						render json: {status: 'ERROR', message:'shoppinglist not saved', data:user.errors},status: :unprocessable_entity
+	 				end
  				else
-					render json: {status: 'ERROR', message:'shoppinglist not saved', data:user.errors},status: :unprocessable_entity
+					render json: {status: 'ERROR', message:'shoppinglist not saved', data:"already has this username"},status: :unprocessable_entity
  				end
 			end
 
-			# use DELETE "http://localhost:3000/api/v1/users/rachel_green2"
+			# use DELETE "http://localhost:3000/api/v1/users/hiby1?password=12345"
 
 			def destroy
-				user = User.find_by_user_name(params[:id])
-				if user
+				user = User.find_by_user_name(params[:user_name])
+				if user && user.authenticate(params[:password])
 					user.destroy
 					render json: {status: 'SUCCESS', message:'Deleted shoppinglist', data:user}, status: :ok
+				elsif user
+					render json: {status: 'ERROR', message:'cannot delete user', data:"username and password are not match, so I cannot delete it."},status: :unprocessable_entity				
 				else
-					render json: {status: 'ERROR', message:'shoppinglist cannot find', data:"cannot find shopping list, so I cannot delete it."},status: :unprocessable_entity
+					render json: {status: 'ERROR', message:'cannot delete user', data:"cannot find shopping list, so I cannot delete it."},status: :unprocessable_entity
 				end
 			end
 
-			# use PUT "http://localhost:3000/api/v1/users/Richard%20Burke"
+			# use PUT "http://localhost:3000/api/v1/users/hiby1?password=12345"
 			# head Content-Type : application/json
 			# body
 			# {
@@ -63,15 +68,17 @@ module Api
 			# }
 
 			def update
-				user = User.find_by_user_name(params[:username])
+				user = User.find_by_user_name(params[:user_name])
 				if user && user.authenticate(params[:password])
 					if user.update_attributes(user_params)
 						render json: {status: 'SUCCESS', message:'Updated shoppinglist', data:user},status: :ok
 					else
-						render json: {status: 'ERROR', message:'shoppinglist not updated', data:user.errors},status: :ok
+						render json: {status: 'ERROR', message:'shoppinglist not updated', data:user.errors},status: :unprocessable_entity
 					end
+				elsif user
+					render json: {status: 'ERROR', message:'shoppinglist not updated', data:"username and password are not match"},status: :unprocessable_entity
 				else
-					render json: {status: 'ERROR', message:'shoppinglist not updated', data:"cannot find user name"},status: :ok
+					render json: {status: 'ERROR', message:'shoppinglist not updated', data:"cannot find user name"},status: :unprocessable_entity
 				end
 			end
